+++ src/index.js
@@ -6,10 +6,85 @@ import 'core-js/fn/array/filter';
 import 'core-js/fn/set';
 import {init} from './lib/init';
 import { CMP_GLOBAL_NAME } from './lib/cmp';
+import log from "./lib/log";
+import config from "./lib/config";
+import { decodeConsentData } from "./lib/cookie/cookie";
+
+function handleConsentResult(cmp, store = {}, vendorList, consentData = {}) {
+	const { isConsentToolShowing } = store;
+
+	const {
+		vendorListVersion: listVersion,
+		tcfPolicyVersion: listPolicyVersion = 1
+	} = vendorList;
+
+	const {
+		created,
+		vendorListVersion,
+		policyVersion: consentPolicyVersion = 1
+	} = consentData;
+
+	if (!created) {
+		log.debug('No consent data found. Showing consent tool');
+		config.autoDisplay && cmp('showConsentTool');
+	} else if (!listVersion) {
+		log.debug('Could not determine vendor list version. Not showing consent tool');
+	} else if (vendorListVersion !== listVersion) {
+		log.debug(`Consent found for version ${vendorListVersion}, but received vendor list version ${listVersion}. Showing consent tool`);
+		config.autoDisplay && cmp('showConsentTool');
+	} else if (consentPolicyVersion !== listPolicyVersion) {
+		log.debug(`Consent found for policy ${consentPolicyVersion}, but received vendor list with policy ${consentPolicyVersion}. Showing consent tool`);
+		config.autoDisplay && cmp('showConsentTool');
+	} else {
+		log.debug('Consent found. Not showing consent tool. Show footer when not all consents set to true');
+		!isConsentToolShowing && config.autoDisplay && cmp('showFooter');
+	}
+}
+
+function checkConsent(cmp, store) {
+	if (!cmp) {
+		log.error('CMP failed to load');
+	}
+	else if (!window.navigator.cookieEnabled) {
+		log.warn('Cookies are disabled. Ignoring CMP consent check');
+	}
+	else {
+		// config.getVendorList
+		__tcfapi('getVendorList', 2,  (vendorList, success) => {
+			if (success) {
+				const timeout = setTimeout(() => {
+					handleConsentResult(cmp, store, vendorList);
+				}, 100);
+
+				// config.getTCData
+				__tcfapi('getTCData', 2, (tcData, success) => {
+					if (success) {
+						let tcStringDecoded;
+
+						try {
+							tcStringDecoded = decodeConsentData(tcData.tcString);
+						} catch (e) {
+							// error ocurred during decoding TCString
+						} finally {
+							clearTimeout(timeout);
+							handleConsentResult(cmp, store, vendorList, tcStringDecoded);
+						}
+					}
+				});
+			}
+		});
+	}
+}
 
 function start() {
-	const {config} = window[CMP_GLOBAL_NAME] || {};
-	init(config);
+	// Preserve any config options already set
+	const { config } = window[CMP_GLOBAL_NAME] || {};
+	const configUpdates = {
+		globalConsentLocation: 'https://rasp.mgr.consensu.org/portal.html',
+		...config
+	};
+
+	init(configUpdates).then((store) => checkConsent(window.__cmp, store));
 }
 
 start();
