+++ src/lib/init.js
@@ -2,7 +2,6 @@ import { h, render } from 'preact';
 import Promise from 'promise-polyfill';
 import Store from './store';
 import Cmp, { CMP_GLOBAL_NAME } from './cmp';
-import { TCString } from "@iabtcf/core";
 import { CmpApi } from '@iabtcf/cmpapi';
 import { fetchGlobalVendorList } from './vendor';
 import { decodeConsentData, readConsentCookie } from './cookie/cookie';
@@ -14,57 +13,6 @@ import createCommands from "./commands";
 const CMP_VERSION = 2;
 const CMP_ID = 280;
 const COOKIE_VERSION = 2;
-
-function handleConsentResult(cmp, {isConsentToolShowing}, {vendorListVersion: listVersion, tcfPolicyVersion: listPolicyVersion = 1} = {}, {created, vendorListVersion, policyVersion: consentPolicyVersion = 1} = {}) {
-	if (!created) {
-		log.debug('No consent data found. Showing consent tool');
-		config.autoDisplay && cmp('showConsentTool');
-	} else if (!listVersion) {
-		log.debug('Could not determine vendor list version. Not showing consent tool');
-	} else if (vendorListVersion !== listVersion) {
-		log.debug(`Consent found for version ${vendorListVersion}, but received vendor list version ${listVersion}. Showing consent tool`);
-		config.autoDisplay && cmp('showConsentTool');
-	} else if (consentPolicyVersion !== listPolicyVersion) {
-		log.debug(`Consent found for policy ${consentPolicyVersion}, but received vendor list with policy ${consentPolicyVersion}. Showing consent tool`);
-		config.autoDisplay && cmp('showConsentTool');
-	} else {
-		log.debug('Consent found. Not showing consent tool. Show footer when not all consents set to true');
-		!isConsentToolShowing && config.autoDisplay && cmp('showFooter');
-	}
-}
-
-function checkConsent(cmp, store) {
-	if (!cmp) {
-		log.error('CMP failed to load');
-	}
-	else if (!window.navigator.cookieEnabled) {
-		log.warn('Cookies are disabled. Ignoring CMP consent check');
-	}
-	else {
-		__tcfapi('getVendorList', 2,  (vendorList, success) => {
-			if (success) {
-				const timeout = setTimeout(() => {
-					handleConsentResult(cmp, store, vendorList);
-				}, 100);
-
-				__tcfapi('getTCData', 2, (tcData, success) => {
-					if (success) {
-						let tcStringDecoded;
-
-						try {
-							tcStringDecoded = TCString.decode(tcData.tcString);
-						} catch (e) {
-							// error ocurred during decoding TCString
-						} finally {
-							clearTimeout(timeout);
-							handleConsentResult(cmp, store, vendorList, tcStringDecoded);
-						}
-					}
-				});
-			}
-		});
-	}
-}
 
 function readExternalConsentData(config) {
 	return new Promise((resolve, reject) => {
@@ -86,7 +34,7 @@ function readExternalConsentData(config) {
 	});
 }
 
-export function init(configUpdates) {
+export function init (configUpdates) {
 	config.update(configUpdates);
 	log.debug('Using configuration:', config);
 
@@ -152,13 +100,3 @@ export function init(configUpdates) {
 			log.error('Failed to load CMP', err);
 		});
 }
-
-// Preserve any config options already set
-const {config: configuration} = window[CMP_GLOBAL_NAME] || {};
-const configUpdates = {
-	globalConsentLocation: 'https://rasp.mgr.consensu.org/portal.html',
-	...configuration
-};
-
-init(configUpdates).then((store) => checkConsent(window.__cmp, store));
-
